Rename metadata_utils to smile_inchi_inchikey_conversions.py
